pyproject.toml
[project]
name = "heim-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Guillermo A. Moran-Arreola",email = "g.a.moran.arreola@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "heim_core", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-pyproject (>=1.2.0,<2.0.0)"
]

[tool.black]
extend-exclude = "migrations"

[tool.isort]
extend_skip = ["migrations"]

[tool.flake8]
max-line-length = 88
extend-exclude = ["migrations"]
